mysite/requirements.txt
backports.csv==1.0.7
defusedxml==0.5.0
diff-match-patch==20181111
Django==1.11.4
django-crispy-forms==1.7.2
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.2.0
django-redis==4.10.0
django-reversion==3.0.3
django-tinymce==2.8.0
DjangoUeditor==1.8.143
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
jdcal==1.4
jieba==0.39
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.0.0
PyMySQL==0.9.3
pytz==2018.9
PyYAML==5.1
redis==3.2.1
six==1.10.0
tablib==0.13.0
Whoosh==2.7.4
xlrd==1.2.0
xlwt==1.3.0
